requirements.txt
streamlit==1.28.0
PyPDF2==3.0.1
openai==1.14.0
tqdm==4.66.1
pdf2image==1.16.3
Pillow==10.0.0
pyngrok==7.0.0
python-dotenv==1.0.0
